# internal/handlers/toddler_handler.go
package handlers

import (
	"grovia/internal/dto/requests"
	"grovia/internal/dto/responses"
	"grovia/internal/services"
	"grovia/pkg"
	"strconv"
	"time"

	"github.com/gofiber/fiber/v2"
)

type ToddlerHandler struct {
	service services.ToddlerService
}

func NewToddlerHandler(service services.ToddlerService) *ToddlerHandler {
	return &ToddlerHandler{service: service}
}

// CreateToddler godoc
// @Summary Create new toddler
// @Description Create a new toddler record
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body requests.CreateToddlerRequest true "Toddler data"
// @Success 201 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Router /toddlers [post]
func (t *ToddlerHandler) CreateToddler(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	var req requests.CreateToddlerRequest
	if err := ctx.BodyParser(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	req.LocationID = locationID

	toddlerResp, predictResponse, err := t.service.CreateToddler(req, userID)
	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusCreated).JSON(responses.BaseResponse{
		Success: true,
		Message: "Create Toddler Success",
		Data: fiber.Map{
			"toddler": toddlerResp,
			"predict": predictResponse,
		},
		Error: nil,
	})
}

// CreateToddlerWithParent godoc
// @Summary Create toddler with parent
// @Description Create a new toddler record along with parent data
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body requests.CreateToddlerWithParentRequest true "Toddler and parent data"
// @Success 201 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Router /toddlers/with-parent [post]
func (t *ToddlerHandler) CreateToddlerWithParent(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	var req requests.CreateToddlerWithParentRequest

	if err := ctx.BodyParser(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	req.Toddler.LocationID = locationID
	req.Parent.LocationID = locationID

	toddlerResponse, parentResp, predictResponse, err := t.service.CreateToddlerWithParent(req.Toddler, req.Parent, userID)
	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusCreated).JSON(responses.BaseResponse{
		Success: true,
		Message: "Create Parent with Toddler Success",
		Data: fiber.Map{
			"toddler": toddlerResponse,
			"parent":  parentResp,
			"predict": predictResponse,
		},
		Error: nil,
	})
}

// GetAllToddler godoc
// @Summary Get all toddlers
// @Description Get list of all toddlers in location
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param name query string false "Search by name"
// @Param page query int false "Page number"
// @Param limit query int false "Items per page"
// @Success 200 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Router /toddlers [get]
func (t *ToddlerHandler) GetAllToddler(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	name := ctx.Query("name")
	pageStr := ctx.Query("page")
	limitStr := ctx.Query("limit")
	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	toddlers, meta, err := t.service.GetAllToddler(locationID, name, pageStr, limitStr)

	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Get All Parent Data Success",
		Data:    toddlers,
		Meta:    meta,
		Error:   nil,
	})
}

// GetToddlerByID godoc
// @Summary Get toddler by ID
// @Description Get toddler details by ID
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Toddler ID"
// @Success 200 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Failure 404 {object} responses.BaseResponse
// @Router /toddlers/{id} [get]
func (t *ToddlerHandler) GetToddlerByID(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	idParam := ctx.Params("id")
	id, err := strconv.Atoi(idParam)

	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	toddler, err := t.service.GetToddlerByID(id, locationID)

	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Get toddler Data Success",
		Data:    toddler,
		Error:   nil,
	})
}

// UpdateToddlerByID godoc
// @Summary Update toddler by ID
// @Description Update toddler data by ID with prediction
// @Tags Toddlers
// @Accept multipart/form-data
// @Produce json
// @Security BearerAuth
// @Param id path int true "Toddler ID"
// @Param name formData string false "Toddler name"
// @Param birthdate formData string false "Birthdate (RFC3339)"
// @Param profilePicture formData file false "Profile picture"
// @Success 200 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Failure 404 {object} responses.BaseResponse
// @Router /toddlers/{id} [patch]
func (t *ToddlerHandler) UpdateToddlerByID(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	var req requests.UpdateToddlerRequest
	if err := ctx.BodyParser(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	birthdateStr := ctx.FormValue("birthdate")
	if birthdateStr != "" {
		parsedTime, err := time.Parse(time.RFC3339, birthdateStr)
		if err != nil {
			return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
				Success: false,
				Message: "Invalid birthdate format",
				Data:    nil,
				Error: responses.ErrorResponse{
					Code:    "INVALID_REQUEST",
					Message: err.Error(),
				},
			})
		}
		req.Birthdate = &parsedTime
	}

	file, err := ctx.FormFile("profilePicture")
	if err == nil {
		req.ProfilePicture = file
	}

	idParam := ctx.Params("id")
	id, err := strconv.Atoi(idParam)

	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	toddlerResponse, predictResponse, err := t.service.UpdateToddlerByID(ctx.Context(), id, locationID, userID, req)

	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Update Toddler Data Success",
		Data: fiber.Map{
			"toddler": toddlerResponse,
			"predict": predictResponse,
		},
		Error: nil,
	})
}

// UpdateToddlerByIDWithoutPredict godoc
// @Summary Update toddler without prediction
// @Description Update toddler data by ID without triggering prediction
// @Tags Toddlers
// @Accept multipart/form-data
// @Produce json
// @Security BearerAuth
// @Param id path int true "Toddler ID"
// @Param name formData string false "Toddler name"
// @Param birthdate formData string false "Birthdate (RFC3339)"
// @Param profilePicture formData file false "Profile picture"
// @Success 200 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Failure 404 {object} responses.BaseResponse
// @Router /toddlers/{id}/without-predict [patch]
func (t *ToddlerHandler) UpdateToddlerByIDWithoutPredict(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	var req requests.UpdateToddlerRequest
	if err := ctx.BodyParser(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	birthdateStr := ctx.FormValue("birthdate")
	if birthdateStr != "" {
		parsedTime, err := time.Parse(time.RFC3339, birthdateStr)
		if err != nil {
			return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
				Success: false,
				Message: "Invalid birthdate format",
				Data:    nil,
				Error: responses.ErrorResponse{
					Code:    "INVALID_REQUEST",
					Message: err.Error(),
				},
			})
		}
		req.Birthdate = &parsedTime
	}

	file, err := ctx.FormFile("profilePicture")
	if err == nil {
		req.ProfilePicture = file
	}

	idParam := ctx.Params("id")
	id, err := strconv.Atoi(idParam)

	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	toddlerResponse, err := t.service.UpdateToddlerByIDWithoutPredict(ctx.Context(), id, locationID, userID, req)

	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Update Toddler Data Success",
		Data:    toddlerResponse,
		Error:   nil,
	})
}

// DeleteToddlerByID godoc
// @Summary Delete toddler by ID
// @Description Delete toddler by ID
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Toddler ID"
// @Success 200 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Failure 404 {object} responses.BaseResponse
// @Router /toddlers/{id} [delete]
func (t *ToddlerHandler) DeleteToddlerByID(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	locationID := ctx.Locals("location_id").(int)

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	idParam := ctx.Params("id")
	id, err := strconv.Atoi(idParam)

	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: err.Error(),
			},
		})
	}

	err = t.service.DeleteToddlerByID(id, locationID, userID)

	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Delete Toddler Data Success",
		Data:    nil,
		Error:   nil,
	})
}

// CheckToddlerExists godoc
// @Summary Check if toddler exists
// @Description Check if toddler exists by name and phone number
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param name query string true "Toddler name"
// @Param phone_number query string true "Parent phone number"
// @Success 200 {object} responses.BaseResponse
// @Failure 400 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Router /toddlers/check [get]
func (t *ToddlerHandler) CheckToddlerExists(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)

	name := ctx.Query("name")
	phoneNumber := ctx.Query("phone_number")

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	if name == "" || phoneNumber == "" {
		return ctx.Status(fiber.StatusBadRequest).JSON(responses.BaseResponse{
			Success: false,
			Message: "Invalid Request",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "INVALID_REQUEST",
				Message: "name is required",
			},
		})
	}

	exists, toddler, err := t.service.CheckToddlerExists(phoneNumber, name)
	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Check Toddler Exists Success",
		Data: fiber.Map{
			"exists": exists,
			"toddler_id": func() *int {
				if toddler != nil {
					return &toddler.ID
				}
				return nil
			}(),
		},
		Error: nil,
	})
}

// GetAllToddlerAllLocation godoc
// @Summary Get all toddlers from all locations
// @Description Get list of all toddlers across all locations (Admin only)
// @Tags Toddlers
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param name query string false "Search by name"
// @Param page query int false "Page number"
// @Param limit query int false "Items per page"
// @Success 200 {object} responses.BaseResponse
// @Failure 401 {object} responses.BaseResponse
// @Router /toddlers/all-locations [get]
func (t *ToddlerHandler) GetAllToddlerAllLocation(ctx *fiber.Ctx) error {
	userID, ok := ctx.Locals("user_id").(int)
	name := ctx.Query("name")
	pageStr := ctx.Query("page")
	limitStr := ctx.Query("limit")

	if !ok || userID == 0 {
		return ctx.Status(fiber.StatusUnauthorized).JSON(responses.BaseResponse{
			Success: false,
			Message: "Unauthorized",
			Data:    nil,
			Error: responses.ErrorResponse{
				Code:    "UNAUTHORIZED",
				Message: "Unauthorized",
			},
		})
	}

	toddlerResponses, meta, err := t.service.GetAllToddlerAllLocation(name, pageStr, limitStr)

	if err != nil {
		return pkg.HandleServiceError(ctx, err)
	}

	return ctx.Status(fiber.StatusOK).JSON(responses.BaseResponse{
		Success: true,
		Message: "Get All Toddler Data Without Location Success",
		Data:    toddlerResponses,
		Meta:    meta,
		Error:   nil,
	})
}